+++ src/Strategy/Maven.hs
@@ -8,18 +8,21 @@ module Strategy.Maven (
 import App.Fossa.Analyze.Types (AnalyzeProject, analyzeProject)
 import App.Pathfinder.Types (LicenseAnalyzeProject, licenseAnalyzeProject)
 import Control.Algebra (Has)
-import Control.Effect.Diagnostics (Diagnostics, context, (<||>))
+import Control.Effect.Diagnostics (Diagnostics, context, warnOnErr, (<||>))
 import Control.Effect.Lift (Lift)
 import Data.Aeson (ToJSON)
+import Diag.Common (MissingDeepDeps (MissingDeepDeps), MissingEdges (MissingEdges))
 import Effect.Exec (Exec)
 import Effect.ReadFS (ReadFS)
 import GHC.Generics (Generic)
+import Graphing (Graphing)
 import Path (Abs, Dir, Path, parent)
 import Strategy.Maven.DepTree qualified as DepTreeCmd
 import Strategy.Maven.PluginStrategy qualified as Plugin
 import Strategy.Maven.Pom qualified as Pom
+import Strategy.Maven.Pom.Closure (MavenProjectClosure)
 import Strategy.Maven.Pom.Closure qualified as PomClosure
-import Types (DependencyResults (..), DiscoveredProject (..), DiscoveredProjectType (MavenProjectType), GraphBreadth (..))
+import Types (Dependency, DependencyResults (..), DiscoveredProject (..), DiscoveredProjectType (MavenProjectType), GraphBreadth (..))
 
 discover ::
   ( Has (Lift IO) sig m
@@ -65,14 +68,54 @@ getDeps ::
   MavenProject ->
   m DependencyResults
 getDeps (MavenProject closure) = do
-  (graph, graphBreadth) <-
-    context "Maven" $
-      context "Plugin analysis" (Plugin.analyze' . parent $ PomClosure.closurePath closure)
-        <||> context "Dynamic analysis" (DepTreeCmd.analyze . parent $ PomClosure.closurePath closure)
-        <||> context "Static analysis" (pure (Pom.analyze' closure, Partial))
+  (graph, graphBreadth) <- context "Maven" $ getDepsDynamicAnalysis closure <||> getStaticAnalysis closure
   pure $
     DependencyResults
       { dependencyGraph = graph
       , dependencyGraphBreadth = graphBreadth
       , dependencyManifestFiles = [PomClosure.closurePath closure]
       }
+
+getDepsDynamicAnalysis ::
+  ( Has (Lift IO) sig m
+  , Has Diagnostics sig m
+  , Has ReadFS sig m
+  , Has Exec sig m
+  ) =>
+  MavenProjectClosure ->
+  m (Graphing Dependency, GraphBreadth)
+getDepsDynamicAnalysis closure =
+  context "Dynamic Analysis" $
+    warnOnErr MissingEdges
+      . warnOnErr MissingDeepDeps
+      $ (getDepsPlugin closure <||> getDepsTreeCmd closure)
+
+getDepsPlugin ::
+  ( Has (Lift IO) sig m
+  , Has Diagnostics sig m
+  , Has ReadFS sig m
+  , Has Exec sig m
+  ) =>
+  MavenProjectClosure ->
+  m (Graphing Dependency, GraphBreadth)
+getDepsPlugin closure = context "Plugin analysis" (Plugin.analyze' . parent $ PomClosure.closurePath closure)
+
+getDepsTreeCmd ::
+  ( Has (Lift IO) sig m
+  , Has Diagnostics sig m
+  , Has ReadFS sig m
+  , Has Exec sig m
+  ) =>
+  MavenProjectClosure ->
+  m (Graphing Dependency, GraphBreadth)
+getDepsTreeCmd closure =
+  context "Dynamic analysis" $
+    DepTreeCmd.analyze . parent $ PomClosure.closurePath closure
+
+getStaticAnalysis ::
+  ( Has (Lift IO) sig m
+  , Has Diagnostics sig m
+  ) =>
+  MavenProjectClosure ->
+  m (Graphing Dependency, GraphBreadth)
+getStaticAnalysis closure = context "Static analysis" $ pure (Pom.analyze' closure, Partial)